add irb120_cam_cali and test folder for the camera
